// src/api/generated/types/MovieResponse.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type MovieResponse = {
    /**
     * @type integer | undefined, int64
    */
    id?: number;
    /**
     * @type integer | undefined, int32
    */
    year?: number;
    /**
     * @type string | undefined
    */
    title?: string;
    /**
     * @type array | undefined
    */
    studios?: string[];
    /**
     * @type array | undefined
    */
    producers?: string[];
    /**
     * @type boolean | undefined
    */
    winner?: boolean;
};